@@ -21,7 +21,7 @@ py_modules = markdown_code_blocks
 install_requires =
     mistune>=2,<3
     pygments
-python_requires = >=3.8
+python_requires = >=3.9
 
 [options.entry_points]
 console_scripts =
